requirements.txt
Flask
pymongo
pycryptodome
bcrypt
flask-jwt-extended
flask-limiter
marshmallow
requests